Initial commit:
This isn't the first commit, but a rebase of about 70 commits which I didnt want
to push all at once. The game is not complete, but I am pushing it here for safe
keeping.

I hope to add a release page soon so that I can upload some build artifacts, perhaps
some native linux/windows binaries.